Prototype of LowFive interface
Attempt to build a C++ interface for building a vol plugin.
Demonstrated with examples/example1.  Uses a generic C plugin
(src/tom-passthru-plugin.c) that captures all the vol calls and routes
them to a custom plugin in C++ in examples/example1/example1_vol.hpp.

Things that I'd like to improve in the design:
- Long list of C wrappers at the bottom of example1_vol.hpp; it would be
better to be in the generic code so the user does not have to copy this, but
they refer to the user's Vol object. Also, them C function names have to
match those in tom-passthru-plugin.c, and the C++ method names have to
match those in VolBase, which is fragile.