WIP: LowFive interface
- Changed src/tom-passthru-plugin from .c and .h to .cpp and .hpp
(added -Wno-varargs to CMakeLists.txt to quiet one warning). Changing
the generic plugin to C++ makes it possible to templatize the callback
functions and easier to call the custom plugin
- Changed examples/example1/example1_vol.hpp to only require
those functions that are being overridden, with no other wrappers
needed.
- Changed src/vol_base.hpp to flatten the VolBase struct, no more nested
structs.
- TODO: figure our how to pass the Vol object to the template parameter
in src/tom-passthru-plugin.cpp. I added a pointer to the derived vol
object to OUR_pass_through_info_t in src/tom-passthru-plugin.hpp and to
VOL_Property in example1/example1.hpp, but I'm not using it for anything
yet. Instead, I'm #including example1_vol.hpp in
tom-passthru-plugin.cpp, which isn't right.