Fix callback-related segfault
a) Add wrappers for VOL classes and expose them instead of directly
exposing DistMetadataVOL, etc
b) In the __init__.py. register a cleanup function with atexit.
It will unset callbacks of the vol.

The issue is that callback members of MetatadataVOL should not be
set or destroyed without GIL acquired, if they come from Python.
Python wrapper allows us to acquire GIL in set_* functions
without interfering with the pure C++ part.